+++ routes/src/actionHistoryRoute.js
@@ -2,6 +2,7 @@ const express = require("express");
 const router = express.Router();
 const authMiddlewares = require("../../middlewares/authMiddlewares");
 const actionHistory = require("../../controllers/actionHistoryController");
+const checkPermission = require("../../middlewares/checkPermission");
 
 // Apply ensureLoggedIn middleware to all routes
 router.use(authMiddlewares.ensureLoggedIn);
@@ -17,11 +18,13 @@ router.get(
 router.post(
   "/nhat-ky-hoat-dong/:id",
   authMiddlewares.ensureRoles(["Admin", "Quản lý"]),
+  checkPermission("delete"),
   actionHistory.deleteData
 );
 router.delete(
   "/nhat-ky-hoat-dong/deleteAllData",
   authMiddlewares.ensureRoles(["Admin", "Quản lý"]),
+  checkPermission("delete"),
   actionHistory.deleteAllData
 );
